move reth related cli macros to base-reth-cli
