+++ Cargo.toml
@@ -286,10 +286,11 @@ base-comp = { path = "crates/batcher/comp", default-features = false }
 base-jwt = { path = "crates/utilities/jwt" }
 base-health = { path = "crates/utilities/health" }
 base-cli-utils = { path = "crates/utilities/cli" }
-base-runtime = { path = "crates/utilities/runtime", default-features = false }
+base-reth-cli = { path = "crates/utilities/reth-cli" }
 base-test-utils = { path = "crates/utilities/test-utils" }
 base-tx-manager = { path = "crates/utilities/tx-manager" }
 base-macros = { path = "crates/utilities/macros", default-features = false }
+base-runtime = { path = "crates/utilities/runtime", default-features = false }
 
 # revm
 revm = { version = "34.0.0", default-features = false }
