+++ bin/builder/Cargo.toml
@@ -17,13 +17,14 @@ workspace = true
 
 [dependencies]
 # workspace
+base-cli-utils.workspace = true
 base-txpool-rpc.workspace = true
 base-node-runner.workspace = true
 base-builder-core.workspace = true
 base-builder-metering.workspace = true
-base-cli-utils = { workspace = true, features = ["reth"] }
 
 # reth
+base-reth-cli.workspace = true
 base-node-core.workspace = true
 reth-cli-util.workspace = true
 base-execution-cli = { workspace = true, features = ["otlp"] }
