+++ bin/node/Cargo.toml
@@ -18,6 +18,7 @@ workspace = true
 [dependencies]
 # workspace
 base-metering.workspace = true
+base-cli-utils.workspace = true
 base-txpool-rpc.workspace = true
 base-flashblocks.workspace = true
 base-node-runner.workspace = true
@@ -26,9 +27,9 @@ base-txpool-tracing.workspace = true
 base-flashblocks-node.workspace = true
 base-proofs-extension.workspace = true
 base-bundle-extension.workspace = true
-base-cli-utils = { workspace = true, features = ["reth"] }
 
 # reth
+base-reth-cli.workspace = true
 base-node-core.workspace = true
 reth-cli-util.workspace = true
 base-execution-cli = { workspace = true, features = ["otlp"] }
