+++ bin/node/src/main.rs
@@ -23,7 +23,7 @@ static ALLOC: reth_cli_util::allocator::Allocator = reth_cli_util::allocator::ne
 
 fn main() {
     base_cli_utils::init_common!();
-    base_cli_utils::init_reth!();
+    base_reth_cli::init_reth!();
 
     let cli = base_cli_utils::parse_cli!(NodeCli);
