+++ crates/utilities/reth-cli/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "base-reth-cli"
+description = "Reth CLI Utilities"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+reth-node-core.workspace = true
